Revert "revert: Move session replay to new consumers (#31490) (#31646)"
This reverts commit 919d07d8639147abf0aa57df96306bb380b52f03.
