+++ plugin-server/src/kafka/consumer.ts
@@ -1,166 +1,53 @@
 import {
-    Assignment,
     ClientMetrics,
-    CODES,
     ConsumerGlobalConfig,
-    ConsumerTopicConfig,
     KafkaConsumer as RdKafkaConsumer,
     LibrdKafkaError,
     Message,
-    TopicPartition,
+    Metadata,
+    PartitionMetadata,
     TopicPartitionOffset,
+    WatermarkOffsets,
 } from 'node-rdkafka'
-import { Gauge } from 'prom-client'
-import { exponentialBuckets } from 'prom-client'
-import { Histogram } from 'prom-client'
+import { hostname } from 'os'
+import { Gauge, Histogram } from 'prom-client'
 
-import { kafkaRebalancePartitionCount, latestOffsetTimestampGauge } from '../main/ingestion-queues/metrics'
+import { defaultConfig } from '../config/config'
 import { logger } from '../utils/logger'
+import { captureException } from '../utils/posthog'
+import { retryIfRetriable } from '../utils/retries'
+import { promisifyCallback } from '../utils/utils'
+import { ensureTopicExists } from './admin'
+import { getConsumerConfigFromEnv } from './config'
 
-export const createKafkaConsumer = async (config: ConsumerGlobalConfig, topicConfig: ConsumerTopicConfig = {}) => {
-    // Creates a node-rdkafka consumer and connects it to the brokers, resolving
-    // only when the connection is established.
+const DEFAULT_BATCH_TIMEOUT_MS = 500
+const DEFAULT_FETCH_BATCH_SIZE = 1000
+const SLOW_BATCH_PROCESSING_LOG_THRESHOLD_MS = 10000
+const MAX_HEALTH_HEARTBEAT_INTERVAL_MS = 60_000
 
-    return await new Promise<RdKafkaConsumer>((resolve, reject) => {
-        const consumer = new RdKafkaConsumer(config, topicConfig)
+const consumedBatchDuration = new Histogram({
+    name: 'consumed_batch_duration_ms',
+    help: 'Main loop consumer batch processing duration in ms',
+    labelNames: ['topic', 'groupId'],
+})
 
-        consumer.on('event.log', (log) => {
-            logger.info('📝', 'librdkafka log', { log: log })
-        })
+const gaugeBatchUtilization = new Gauge({
+    name: 'consumer_batch_utilization',
+    help: 'Indicates how big batches are we are processing compared to the max batch size. Useful as a scaling metric',
+    labelNames: ['groupId'],
+})
 
-        consumer.on('event.error', (error: LibrdKafkaError) => {
-            logger.error('📝', 'librdkafka error', { log: error })
-        })
+const histogramKafkaBatchSize = new Histogram({
+    name: 'consumer_batch_size',
+    help: 'The size of the batches we are receiving from Kafka',
+    buckets: [0, 50, 100, 250, 500, 750, 1000, 1500, 2000, 3000, Infinity],
+})
 
-        consumer.on('subscribed', (topics) => {
-            logger.info('📝', 'librdkafka consumer subscribed', { topics })
-        })
-
-        consumer.on('connection.failure', (error: LibrdKafkaError, metrics: ClientMetrics) => {
-            logger.error('📝', 'librdkafka connection failure', { error, metrics })
-        })
-
-        consumer.on('offset.commit', (error: LibrdKafkaError, topicPartitionOffsets: TopicPartitionOffset[]) => {
-            if (error) {
-                logger.warn('📝', 'librdkafka_offet_commit_error', { error, topicPartitionOffsets })
-            } else {
-                logger.debug('📝', 'librdkafka_offset_commit', { topicPartitionOffsets })
-            }
-        })
-
-        consumer.connect({}, (error, data) => {
-            if (error) {
-                logger.error('⚠️', 'connect_error', { error: error })
-                reject(error)
-            } else {
-                logger.info('📝', 'librdkafka consumer connected', { brokers: data?.brokers })
-                resolve(consumer)
-            }
-        })
-    })
-}
-
-export function countPartitionsPerTopic(assignments: Assignment[]): Map<string, number> {
-    const partitionsPerTopic = new Map()
-    for (const assignment of assignments) {
-        if (partitionsPerTopic.has(assignment.topic)) {
-            partitionsPerTopic.set(assignment.topic, partitionsPerTopic.get(assignment.topic) + 1)
-        } else {
-            partitionsPerTopic.set(assignment.topic, 1)
-        }
-    }
-    return partitionsPerTopic
-}
-
-export const instrumentConsumerMetrics = (consumer: RdKafkaConsumer, groupId: string) => {
-    // For each message consumed, we record the latest timestamp processed for
-    // each partition assigned to this consumer group member. This consumer
-    // should only provide metrics for the partitions that are assigned to it,
-    // so we need to make sure we don't publish any metrics for other
-    // partitions, otherwise we can end up with ghost readings.
-    //
-    // We also need to consider the case where we have a partition that
-    // has reached EOF, in which case we want to record the current time
-    // as opposed to the timestamp of the current message (as in this
-    // case, no such message exists).
-    //
-    // Further, we are not guaranteed to have messages from all of the
-    // partitions assigned to this consumer group member, event if there
-    // are partitions with messages to be consumed. This is because
-    // librdkafka will only fetch messages from a partition if there is
-    // space in the internal partition queue. If the queue is full, it
-    // will not fetch any more messages from the given partition.
-    //
-    // Note that we don't try to align the timestamps with the actual broker
-    // committed offsets. The discrepancy is hopefully in most cases quite
-    // small.
-    //
-    // TODO: add other relevant metrics here
-    // TODO: expose the internal librdkafka metrics as well.
-    consumer.on('rebalance', (error: LibrdKafkaError, assignments: TopicPartition[]) => {
-        /**
-         * see https://github.com/Blizzard/node-rdkafka#rebalancing errors are used to signal
-         * both errors and _not_ errors
-         *
-         * When rebalancing starts the consumer receives ERR_REVOKED_PARTITIONS
-         * And when the balancing is completed the new assignments are received with ERR__ASSIGN_PARTITIONS
-         */
-        if (error.code === CODES.ERRORS.ERR__ASSIGN_PARTITIONS) {
-            logger.info('📝️', `librdkafka cooperative rebalance, partitions assigned`, { assignments })
-            for (const [topic, count] of countPartitionsPerTopic(assignments)) {
-                kafkaRebalancePartitionCount.labels({ topic: topic }).inc(count)
-            }
-        } else if (error.code === CODES.ERRORS.ERR__REVOKE_PARTITIONS) {
-            logger.info('📝️', `librdkafka cooperative rebalance started, partitions revoked`, {
-                revocations: assignments,
-            })
-            for (const [topic, count] of countPartitionsPerTopic(assignments)) {
-                kafkaRebalancePartitionCount.labels({ topic: topic }).dec(count)
-            }
-        } else {
-            // We had a "real" error
-            logger.error('⚠️', 'rebalance_error', { error })
-        }
-
-        latestOffsetTimestampGauge.reset()
-    })
-
-    consumer.on('partition.eof', (topicPartitionOffset: TopicPartitionOffset) => {
-        latestOffsetTimestampGauge
-            .labels({
-                topic: topicPartitionOffset.topic,
-                partition: topicPartitionOffset.partition.toString(),
-                groupId,
-            })
-            .set(Date.now())
-    })
-
-    consumer.on('data', (message) => {
-        if (message.timestamp) {
-            latestOffsetTimestampGauge
-                .labels({ topic: message.topic, partition: message.partition, groupId })
-                .set(message.timestamp)
-        }
-    })
-}
-export const consumeMessages = async (consumer: RdKafkaConsumer, fetchBatchSize: number) => {
-    // Rather than using the pure streaming method of consuming, we
-    // instead fetch in batches. This is to make the logic a little
-    // simpler to start with, although we may want to move to a
-    // streaming implementation if needed. Although given we might want
-    // to switch to a language with better support for Kafka stream
-    // processing, perhaps this will be enough for us.
-    // TODO: handle retriable `LibrdKafkaError`s.
-    return await new Promise<Message[]>((resolve, reject) => {
-        consumer.consume(fetchBatchSize, (error: LibrdKafkaError, messages: Message[]) => {
-            if (error) {
-                reject(error)
-            } else {
-                resolve(messages)
-            }
-        })
-    })
-}
+const histogramKafkaBatchSizeKb = new Histogram({
+    name: 'consumer_batch_size_kb',
+    help: 'The size in kb of the batches we are receiving from Kafka',
+    buckets: [0, 128, 512, 1024, 5120, 10240, 20480, 51200, 102400, 204800, Infinity],
+})
 
 export const findOffsetsToCommit = (messages: TopicPartitionOffset[]): TopicPartitionOffset[] => {
     // We only need to commit the highest offset for a batch of messages
@@ -194,69 +81,296 @@ export const findOffsetsToCommit = (messages: TopicPartitionOffset[]): TopicPart
     return highestOffsets
 }
 
-/**
- * Updates the offsets that will be committed on the next call to commit() (without offsets
- * specified) or the next auto commit.
- *
- * This is a local (in-memory) operation and does not talk to the Kafka broker.
- */
-export const storeOffsetsForMessages = (messages: Message[], consumer: RdKafkaConsumer) => {
-    const topicPartitionOffsets = findOffsetsToCommit(messages).map((message) => {
-        return {
-            ...message,
-            // When committing to Kafka you commit the offset of the next message you want to consume
-            offset: message.offset + 1,
-        }
-    })
-
-    if (topicPartitionOffsets.length > 0) {
-        logger.debug('📝', 'Storing offsets', { topicPartitionOffsets })
-        consumer.offsetsStore(topicPartitionOffsets)
-    }
+export type KafkaConsumerConfig = {
+    groupId: string
+    topic: string
+    batchTimeoutMs?: number
+    callEachBatchWhenEmpty?: boolean
+    autoOffsetStore?: boolean
+    autoCommit?: boolean
 }
 
-export const disconnectConsumer = async (consumer: RdKafkaConsumer) => {
-    await new Promise((resolve, reject) => {
-        consumer.disconnect((error, data) => {
+export type RdKafkaConsumerConfig = Omit<
+    ConsumerGlobalConfig,
+    'group.id' | 'enable.auto.offset.store' | 'enable.auto.commit'
+>
+
+export class KafkaConsumer {
+    private isStopping = false
+    private lastHeartbeatTime = 0
+    private rdKafkaConsumer: RdKafkaConsumer
+    private consumerConfig: ConsumerGlobalConfig
+    private fetchBatchSize: number
+    private maxHealthHeartbeatIntervalMs: number
+    private consumerLoop: Promise<void> | undefined
+
+    constructor(private config: KafkaConsumerConfig, rdKafkaConfig: RdKafkaConsumerConfig = {}) {
+        this.config.autoCommit ??= true
+        this.config.autoOffsetStore ??= true
+        this.config.callEachBatchWhenEmpty ??= false
+
+        this.fetchBatchSize = defaultConfig.CONSUMER_BATCH_SIZE || DEFAULT_FETCH_BATCH_SIZE
+        this.maxHealthHeartbeatIntervalMs =
+            defaultConfig.CONSUMER_MAX_HEARTBEAT_INTERVAL_MS || MAX_HEALTH_HEARTBEAT_INTERVAL_MS
+
+        this.consumerConfig = {
+            'client.id': hostname(),
+            'security.protocol': 'plaintext',
+            'metadata.broker.list': 'kafka:9092', // Overridden with KAFKA_CONSUMER_METADATA_BROKER_LIST
+            log_level: 4, // WARN as the default
+            'group.id': this.config.groupId,
+            'session.timeout.ms': 30_000,
+            'max.poll.interval.ms': 300_000,
+            'max.partition.fetch.bytes': 1_048_576,
+            'fetch.error.backoff.ms': 100,
+            'fetch.message.max.bytes': 10_485_760,
+            'fetch.wait.max.ms': 50,
+            'queued.min.messages': 100000,
+            'queued.max.messages.kbytes': 102400, // 1048576 is the default, we go smaller to reduce mem usage.
+            // Custom settings and overrides - this is where most configuration overrides should be done
+            ...getConsumerConfigFromEnv(),
+            // Finally any specifically given consumer config overrides
+            ...rdKafkaConfig,
+            // Below is config that we explicitly DO NOT want to be overrideable by env vars - i.e. things that would require code changes to change
+            'enable.auto.offset.store': false, // NOTE: This is always false - we handle it using a custom function
+            'enable.auto.commit': this.config.autoCommit,
+            'partition.assignment.strategy': 'cooperative-sticky',
+            'enable.partition.eof': true,
+            rebalance_cb: true,
+            offset_commit_cb: true,
+        }
+
+        this.rdKafkaConsumer = this.createConsumer()
+    }
+
+    public getConfig() {
+        return {
+            ...this.consumerConfig,
+        }
+    }
+
+    public heartbeat() {
+        // Can be called externally to update the heartbeat time and keep the consumer alive
+        this.lastHeartbeatTime = Date.now()
+    }
+
+    public isHealthy() {
+        // this is called as a readiness and a liveness probe
+        const isWithinInterval = Date.now() - this.lastHeartbeatTime < this.maxHealthHeartbeatIntervalMs
+        const isConnected = this.rdKafkaConsumer.isConnected()
+        return isConnected && isWithinInterval
+    }
+
+    public assignments() {
+        return this.rdKafkaConsumer.isConnected() ? this.rdKafkaConsumer.assignments() : []
+    }
+
+    public offsetsStore(topicPartitionOffsets: TopicPartitionOffset[]) {
+        return this.rdKafkaConsumer.offsetsStore(topicPartitionOffsets)
+    }
+
+    public on: RdKafkaConsumer['on'] = (...args) => {
+        // Delegate to the internal consumer
+        return this.rdKafkaConsumer.on(...args)
+    }
+
+    public async queryWatermarkOffsets(topic: string, partition: number, timeout = 10000): Promise<[number, number]> {
+        if (!this.rdKafkaConsumer.isConnected()) {
+            throw new Error('Not connected')
+        }
+
+        const offsets = await promisifyCallback<WatermarkOffsets>((cb) =>
+            this.rdKafkaConsumer.queryWatermarkOffsets(topic, partition, timeout, cb)
+        ).catch((err) => {
+            captureException(err)
+            logger.error('🔥', 'Failed to query kafka watermark offsets', err)
+            throw err
+        })
+
+        return [offsets.lowOffset, offsets.highOffset]
+    }
+
+    public async getPartitionsForTopic(topic: string): Promise<PartitionMetadata[]> {
+        if (!this.rdKafkaConsumer.isConnected()) {
+            throw new Error('Not connected')
+        }
+
+        const meta = await promisifyCallback<Metadata>((cb) => this.rdKafkaConsumer.getMetadata({ topic }, cb)).catch(
+            (err) => {
+                captureException(err)
+                logger.error('🔥', 'Failed to get partition metadata', err)
+                throw err
+            }
+        )
+
+        return meta.topics.find((x) => x.name === topic)?.partitions ?? []
+    }
+
+    private createConsumer() {
+        const consumer = new RdKafkaConsumer(this.consumerConfig, {
+            // Default settings
+            'auto.offset.reset': 'earliest',
+        })
+
+        consumer.on('event.log', (log) => {
+            logger.info('📝', 'librdkafka log', { log: log })
+        })
+
+        consumer.on('event.error', (error: LibrdKafkaError) => {
+            logger.error('📝', 'librdkafka error', { log: error })
+        })
+
+        consumer.on('subscribed', (topics) => {
+            logger.info('📝', 'librdkafka consumer subscribed', { topics, config: this.consumerConfig })
+        })
+
+        consumer.on('connection.failure', (error: LibrdKafkaError, metrics: ClientMetrics) => {
+            logger.error('📝', 'librdkafka connection failure', { error, metrics, config: this.consumerConfig })
+        })
+
+        consumer.on('offset.commit', (error: LibrdKafkaError, topicPartitionOffsets: TopicPartitionOffset[]) => {
             if (error) {
-                logger.error('🔥', 'Failed to disconnect node-rdkafka consumer', { error })
-                reject(error)
+                logger.warn('📝', 'librdkafka_offet_commit_error', { error, topicPartitionOffsets })
             } else {
-                logger.info('🔁', 'Disconnected node-rdkafka consumer')
-                resolve(data)
+                logger.debug('📝', 'librdkafka_offset_commit', { topicPartitionOffsets })
             }
         })
-    })
+
+        return consumer
+    }
+
+    private storeOffsetsForMessages = (messages: Message[]) => {
+        const topicPartitionOffsets = findOffsetsToCommit(messages).map((message) => {
+            return {
+                ...message,
+                // When committing to Kafka you commit the offset of the next message you want to consume
+                offset: message.offset + 1,
+            }
+        })
+
+        if (topicPartitionOffsets.length > 0) {
+            logger.debug('📝', 'Storing offsets', { topicPartitionOffsets })
+            this.rdKafkaConsumer.offsetsStore(topicPartitionOffsets)
+        }
+    }
+
+    public async connect(eachBatch: (messages: Message[]) => Promise<void>) {
+        const { topic, groupId, callEachBatchWhenEmpty = false } = this.config
+
+        try {
+            await promisifyCallback<Metadata>((cb) => this.rdKafkaConsumer.connect({}, cb))
+            logger.info('📝', 'librdkafka consumer connected')
+        } catch (error) {
+            logger.error('⚠️', 'connect_error', { error: error })
+            throw error
+        }
+
+        this.heartbeat() // Setup the heartbeat so we are healthy since connection is established
+
+        await ensureTopicExists(this.consumerConfig, this.config.topic)
+
+        // The consumer has an internal pre-fetching queue that sequentially pools
+        // each partition, with the consumerMaxWaitMs timeout. We want to read big
+        // batches from this queue, but guarantee we are still running (with smaller
+        // batches) if the queue is not full enough. batchingTimeoutMs is that
+        // timeout, to return messages even if fetchBatchSize is not reached.
+        this.rdKafkaConsumer.setDefaultConsumeTimeout(this.config.batchTimeoutMs || DEFAULT_BATCH_TIMEOUT_MS)
+        this.rdKafkaConsumer.subscribe([this.config.topic])
+
+        const startConsuming = async () => {
+            try {
+                while (!this.isStopping) {
+                    logger.debug('🔁', 'main_loop_consuming')
+
+                    // TRICKY: We wrap this in a retry check. It seems that despite being connected and ready, the client can still have an undeterministic
+                    // error when consuming, hence the retryIfRetriable.
+                    const messages = await retryIfRetriable(() =>
+                        promisifyCallback<Message[]>((cb) => this.rdKafkaConsumer.consume(this.fetchBatchSize, cb))
+                    )
+
+                    // After successfully pulling a batch, we can update our heartbeat time
+                    this.heartbeat()
+
+                    gaugeBatchUtilization.labels({ groupId }).set(messages.length / this.fetchBatchSize)
+
+                    logger.debug('🔁', 'main_loop_consumed', { messagesLength: messages.length })
+                    histogramKafkaBatchSize.observe(messages.length)
+                    histogramKafkaBatchSizeKb.observe(
+                        messages.reduce((acc, m) => (m.value?.length ?? 0) + acc, 0) / 1024
+                    )
+
+                    if (!messages.length && !callEachBatchWhenEmpty) {
+                        logger.debug('🔁', 'main_loop_empty_batch', { cause: 'empty' })
+                        continue
+                    }
+
+                    const startProcessingTimeMs = new Date().valueOf()
+                    await eachBatch(messages)
+
+                    const processingTimeMs = new Date().valueOf() - startProcessingTimeMs
+                    consumedBatchDuration.labels({ topic, groupId }).observe(processingTimeMs)
+
+                    const logSummary = `Processed ${messages.length} events in ${
+                        Math.round(processingTimeMs / 10) / 100
+                    }s`
+                    if (processingTimeMs > SLOW_BATCH_PROCESSING_LOG_THRESHOLD_MS) {
+                        logger.warn('🕒', `Slow batch: ${logSummary}`)
+                    }
+
+                    if (this.config.autoCommit && this.config.autoOffsetStore) {
+                        this.storeOffsetsForMessages(messages)
+                    }
+                }
+            } catch (error) {
+                throw error
+            } finally {
+                logger.info('🔁', 'main_loop_stopping')
+
+                // Finally, disconnect from the broker. If stored offsets have changed via
+                // `storeOffsetsForMessages` above, they will be committed before shutdown (so long
+                // as this consumer is still part of the group).
+                await this.disconnectConsumer()
+                logger.info('🔁', 'Disconnected node-rdkafka consumer')
+            }
+        }
+
+        this.consumerLoop = startConsuming().catch((error) => {
+            logger.error('🔁', 'consumer_loop_error', {
+                error: String(error),
+                config: this.config,
+                consumerConfig: this.consumerConfig,
+            })
+            // We re-throw the error as that way it will be caught in server.ts and trigger a full shutdown
+            throw error
+        })
+    }
+
+    public async disconnect() {
+        if (this.isStopping) {
+            return
+        }
+        // Mark as stopping - this will also essentially stop the consumer loop
+        this.isStopping = true
+
+        // Allow the in progress consumer loop to finish if possible
+        if (this.consumerLoop) {
+            await this.consumerLoop.catch((error) => {
+                logger.error('🔁', 'failed to stop consumer loop safely. Continuing shutdown', {
+                    error: String(error),
+                    config: this.config,
+                    consumerConfig: this.consumerConfig,
+                })
+            })
+        }
+
+        await this.disconnectConsumer()
+    }
+
+    private async disconnectConsumer() {
+        if (this.rdKafkaConsumer.isConnected()) {
+            logger.info('📝', 'Disconnecting consumer...')
+            await new Promise<void>((res, rej) => this.rdKafkaConsumer.disconnect((e) => (e ? rej(e) : res())))
+            logger.info('📝', 'Disconnected consumer!')
+        }
+    }
 }
-
-export const consumedBatchDuration = new Histogram({
-    name: 'consumed_batch_duration_ms',
-    help: 'Main loop consumer batch processing duration in ms',
-    labelNames: ['topic', 'groupId'],
-})
-
-export const consumerBatchSize = new Histogram({
-    name: 'consumed_batch_size',
-    help: 'Size of the batch fetched by the consumer',
-    labelNames: ['topic', 'groupId'],
-    buckets: exponentialBuckets(1, 3, 5),
-})
-
-export const consumedMessageSizeBytes = new Histogram({
-    name: 'consumed_message_size_bytes',
-    help: 'Size of consumed message value in bytes',
-    labelNames: ['topic', 'groupId', 'messageType'],
-    buckets: exponentialBuckets(1, 8, 4).map((bucket) => bucket * 1024),
-})
-
-export const kafkaAbsolutePartitionCount = new Gauge({
-    name: 'kafka_absolute_partition_count',
-    help: 'Number of partitions assigned to this consumer. (Absolute value from the consumer state.)',
-    labelNames: ['topic'],
-})
-
-export const gaugeBatchUtilization = new Gauge({
-    name: 'consumer_batch_utilization',
-    help: 'Indicates how big batches are we are processing compared to the max batch size. Useful as a scaling metric',
-    labelNames: ['groupId'],
-})
