+++ plugin-server/src/main/ingestion-queues/session-recording/utils.ts
@@ -1,5 +1,5 @@
 import { DateTime } from 'luxon'
-import { KafkaConsumer, Message, MessageHeader, PartitionMetadata } from 'node-rdkafka'
+import { Message, MessageHeader } from 'node-rdkafka'
 import path from 'path'
 import { Counter } from 'prom-client'
 
@@ -7,7 +7,6 @@ import { KafkaProducerWrapper } from '../../../kafka/producer'
 import { PipelineEvent, RawEventMessage, RRWebEvent } from '../../../types'
 import { parseJSON } from '../../../utils/json-parse'
 import { logger } from '../../../utils/logger'
-import { captureException } from '../../../utils/posthog'
 import { captureIngestionWarning } from '../../../worker/ingestion/utils'
 import { eventDroppedCounter } from '../metrics'
 import { TeamIDWithConfig } from './session-recordings-consumer'
@@ -59,49 +58,6 @@ export const maxDefined = (...args: (number | undefined)[]): number | undefined 
 }
 
 export const bufferFileDir = (root: string) => path.join(root, 'session-buffer-files')
-
-export const queryWatermarkOffsets = (
-    kafkaConsumer: KafkaConsumer | undefined,
-    topic: string,
-    partition: number,
-    timeout = 10000
-): Promise<[number, number]> => {
-    return new Promise<[number, number]>((resolve, reject) => {
-        if (!kafkaConsumer) {
-            return reject('Not connected')
-        }
-
-        kafkaConsumer.queryWatermarkOffsets(topic, partition, timeout, (err, offsets) => {
-            if (err) {
-                captureException(err)
-                logger.error('🔥', 'Failed to query kafka watermark offsets', err)
-                return reject(err)
-            }
-
-            resolve([partition, offsets.highOffset])
-        })
-    })
-}
-
-export const getPartitionsForTopic = (
-    kafkaConsumer: KafkaConsumer | undefined,
-    topic: string
-): Promise<PartitionMetadata[]> => {
-    return new Promise<PartitionMetadata[]>((resolve, reject) => {
-        if (!kafkaConsumer) {
-            return reject('Not connected')
-        }
-        kafkaConsumer.getMetadata({ topic }, (err, meta) => {
-            if (err) {
-                captureException(err)
-                logger.error('🔥', 'Failed to get partition metadata', err)
-                return reject(err)
-            }
-
-            return resolve(meta.topics.find((x) => x.name === topic)?.partitions ?? [])
-        })
-    })
-}
 
 export const getLagMultiplier = (lag: number, threshold = 1000000) => {
     if (lag < threshold) {
